requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiogram==2.25.1
aiogram-utils==0.1.4
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
mongoengine==0.24.2
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
typing_extensions==4.7.1
ujson==5.8.0
yarl==1.9.2
